+++ tests/Extensions/Hosting.Services.Web.UnitTests/WebEndpointInfoTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license.
 
 using System;
+using Microsoft.Omex.Extensions.ServiceFabricGuest.Abstractions;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Microsoft.Omex.Extensions.Hosting.Services.Web.UnitTests.Internal
